Lint: scope npm run lint to supported source and clean mechanical src failures
- npm run lint now targets src, vite.config.ts, and build/test scripts only;
  one-off diagnostic scripts move to the optional lint:scripts command
- Fix mechanical src lint errors: prefer-const, no-self-assign, useless
  escape, statement-position ternaries, stale @ts-ignore directives
- Extract ToolpathVisibility type/default const and UnsupportedMobileScreen
  into their own files to satisfy react-refresh/only-export-components
- Keep the three "unused" react-hooks/exhaustive-deps disable directives:
  removing them unmasks compiler-backed react-hooks errors (documented in
  planning/LINT_HOOK_TYPING_DEBT_Plan.md)
- Add approved design plan for the remaining hook/ref/typing lint debt
- src lint baseline: 95 problems -> 74 (72 errors -> 51)
