+++ src/components/canvas/operationSnapshot.ts
@@ -18,7 +18,7 @@ import { getFeatureGeometryProfiles } from '../../text'
 import { getProfileBounds, rectProfile } from '../../types/project'
 import type { Operation, Project, SketchProfile } from '../../types/project'
 import type { ToolpathResult } from '../../engine/toolpaths'
-import type { ToolpathVisibility } from '../ToolpathVisibilityPanel'
+import type { ToolpathVisibility } from '../toolpathVisibility'
 import { drawToolpath } from './previewPrimitives'
 import { traceProfilePath } from './profilePrimitives'
 import { drawClampFootprint, drawTabFootprint } from './scenePrimitives'
