+++ src/components/toolpathVisibility.ts
@@ -1,0 +1,31 @@
+/**
+ * Copyright 2026 Franja (Frank) Povazanj
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export interface ToolpathVisibility {
+  cuts: boolean
+  rapids: boolean
+  plunges: boolean
+  retractions: boolean
+  directions: boolean
+}
+
+export const DEFAULT_TOOLPATH_VISIBILITY: ToolpathVisibility = {
+  cuts: true,
+  rapids: true,
+  plunges: true,
+  retractions: true,
+  directions: true,
+}
